patch: [N4S] Use self
