+++ packages/n4s/src/enforce/__tests__/enforce.test.js
@@ -1,6 +1,6 @@
 import rules from 'rules';
 
-const allRules = Object.keys(rules);
+const allRules = Object.keys(rules());
 const _proxy = Proxy;
 
 const suite = ({ withProxy, requirePath }) =>
