+++ packages/n4s/src/rules/__tests__/rules.test.js
@@ -2,8 +2,8 @@ import rules from 'rules';
 
 describe('Tests enforce rules API', () => {
   it('Should expose all enforce rules', () => {
-    Object.keys(rules).forEach(rule => {
-      expect(rules[rule]).toBeInstanceOf(Function);
+    Object.keys(rules()).forEach(rule => {
+      expect(rules()[rule]).toBeInstanceOf(Function);
     });
   });
 });
