+++ packages/n4s/src/rules/endsWith.js
@@ -1,11 +1,8 @@
 import bindNot from 'bindNot';
+import isString from 'isStringValue';
 
 export function endsWith(value, arg1) {
-  return (
-    typeof value === 'string' &&
-    typeof arg1 === 'string' &&
-    value.endsWith(arg1)
-  );
+  return isString(value) && isString(arg1) && value.endsWith(arg1);
 }
 
 export const doesNotEndWith = bindNot(endsWith);
