+++ packages/n4s/src/rules/isEmpty.js
@@ -1,5 +1,6 @@
 import bindNot from 'bindNot';
 import { isNumeric } from 'isNumeric';
+import { lengthEquals } from 'lengthEquals';
 
 export function isEmpty(value) {
   if (!value) {
@@ -7,12 +8,12 @@ export function isEmpty(value) {
   } else if (isNumeric(value)) {
     return value === 0;
   } else if (Object.prototype.hasOwnProperty.call(value, 'length')) {
-    return value.length === 0;
+    return lengthEquals(value, 0);
   } else if (typeof value === 'object') {
-    return Object.keys(value).length === 0;
-  } else {
-    return true;
+    return lengthEquals(Object.keys(value), 0);
   }
+
+  return true;
 }
 
 export const isNotEmpty = bindNot(isEmpty);
