+++ packages/n4s/src/rules/matches.js
@@ -1,9 +1,10 @@
 import bindNot from 'bindNot';
+import isString from 'isStringValue';
 
 export function matches(value, regex) {
   if (regex instanceof RegExp) {
     return regex.test(value);
-  } else if (typeof regex === 'string') {
+  } else if (isString(regex)) {
     return new RegExp(regex).test(value);
   } else {
     return false;
