+++ packages/n4s/src/rules/rules.js
@@ -27,55 +27,57 @@ import { shorterThan } from 'shorterThan';
 import { shorterThanOrEquals } from 'shorterThanOrEquals';
 import { startsWith, doesNotStartWith } from 'startsWith';
 
-export default {
-  doesNotEndWith,
-  doesNotStartWith,
-  endsWith,
-  equals,
-  greaterThan,
-  greaterThanOrEquals,
-  gt: greaterThan,
-  gte: greaterThanOrEquals,
-  inside,
-  isArray,
-  isBetween,
-  isEmpty,
-  isEven,
-  isFalsy,
-  isNaN,
-  isNegative,
-  isNotArray,
-  isNotBetween,
-  isNotEmpty,
-  isNotNaN,
-  isNotNull,
-  isNotNumber,
-  isNotNumeric,
-  isNotString,
-  isNotUndefined,
-  isNull,
-  isNumber,
-  isNumeric,
-  isOdd,
-  isPositive,
-  isString,
-  isTruthy,
-  isUndefined,
-  lengthEquals,
-  lengthNotEquals,
-  lessThan,
-  lessThanOrEquals,
-  longerThan,
-  longerThanOrEquals,
-  lt: lessThan,
-  lte: lessThanOrEquals,
-  matches,
-  notEquals,
-  notInside,
-  notMatches,
-  numberEquals,
-  numberNotEquals,
-  shorterThan,
-  shorterThanOrEquals,
-  startsWith,
-};
+export default function rules() {
+  return {
+    doesNotEndWith,
+    doesNotStartWith,
+    endsWith,
+    equals,
+    greaterThan,
+    greaterThanOrEquals,
+    gt: greaterThan,
+    gte: greaterThanOrEquals,
+    inside,
+    isArray,
+    isBetween,
+    isEmpty,
+    isEven,
+    isFalsy,
+    isNaN,
+    isNegative,
+    isNotArray,
+    isNotBetween,
+    isNotEmpty,
+    isNotNaN,
+    isNotNull,
+    isNotNumber,
+    isNotNumeric,
+    isNotString,
+    isNotUndefined,
+    isNull,
+    isNumber,
+    isNumeric,
+    isOdd,
+    isPositive,
+    isString,
+    isTruthy,
+    isUndefined,
+    lengthEquals,
+    lengthNotEquals,
+    lessThan,
+    lessThanOrEquals,
+    longerThan,
+    longerThanOrEquals,
+    lt: lessThan,
+    lte: lessThanOrEquals,
+    matches,
+    notEquals,
+    notInside,
+    notMatches,
+    numberEquals,
+    numberNotEquals,
+    shorterThan,
+    shorterThanOrEquals,
+    startsWith,
+  };
+}
