+++ packages/n4s/src/rules/startsWith.js
@@ -1,11 +1,8 @@
 import bindNot from 'bindNot';
+import isString from 'isStringValue';
 
 export function startsWith(value, arg1) {
-  return (
-    typeof value === 'string' &&
-    typeof arg1 === 'string' &&
-    value.startsWith(arg1)
-  );
+  return isString(value) && isString(arg1) && value.startsWith(arg1);
 }
 
 export const doesNotStartWith = bindNot(startsWith);
