+++ packages/vest/src/core/test/lib/addTestToState.js
@@ -1,0 +1,9 @@
+import useTestObjects from 'useTestObjects';
+
+/**
+ * Stores test object inside suite state.
+ * @param {VestTest} testObject
+ */
+export default testObject => {
+  useTestObjects(testObjects => testObjects.concat(testObject));
+};
